+++ cookbooks/maxlab_tftp_server/metadata.rb
@@ -1,25 +1,26 @@
-name 'maxlab_pxe'
+name 'maxlab_tftp_server'
 maintainer 'Maxwell Spangler'
 maintainer_email 'maxcode@maxwellspangler.com'
 license 'All Rights Reserved'
-description 'Deploys infrastructure files for PXE booting'
-long_description 'Deploys infrastructure files for PXE booting'
-version '0.1.1'
+description 'Deploys a TFTP server for the maxlab home lab'
+long_description 'Deploys a TFTP server for the maxlab home lab'
+version '0.1.0'
 chef_version '>= 13.0'
 
 # The `issues_url` points to the location where issues for this cookbook are
 # tracked.  A `View Issues` link will be displayed on this cookbook's page when
 # uploaded to a Supermarket.
 #
-# issues_url 'https://github.com/<insert_org_here>/maxlab_pxe/issues'
+# issues_url 'https://github.com/<insert_org_here>/maxlab-tftp-server/issues'
 
 # The `source_url` points to the development repository for this cookbook.  A
 # `View Source` link will be displayed on this cookbook's page when uploaded to
 # a Supermarket.
 #
-# source_url 'https://github.com/<insert_org_here>/maxlab_pxe'
+# source_url 'https://github.com/<insert_org_here>/maxlab-tftp-server'
 
-supports 'redhat', '>= 7.0'
-supports 'centos', '>= 7.0'
-supports 'oracle', '>= 7.0'
-supports 'fedora', '>= 30.0'
+supports 'redhat', '>= 8.0'
+supports 'centos', '>= 8.0'
+supports 'oracle', '>= 8.0'
+
+depends 'maxlab_firewall'