+++ cookbooks/maxlab_tftp_server/recipes/default.rb
@@ -1,0 +1,12 @@
+#
+# Cookbook:: maxlab-tftp
+# Recipe:: default
+#
+# Copyright:: 2019, The Authors, All Rights Reserved.
+#
+
+=begin
+#<
+The default recipe performs no actions.
+#>
+=end