+++ cookbooks/maxlab_pxe/CHANGELOG.md
@@ -1,5 +1,10 @@
 # maxlab_pxe CHANGELOG
 
+# 0.1.1
+
+* Added recipe images.rb - creates and populates /var/lib/tftpboot/images with Linux kernel and initrd files as well as netboot tools like Clonezilla
+* Added recipe syslinux - installs syslinux PXE menu files
+
 # 0.1.0
 
 * Draft release